confirm checking voice

+++ src/handlers/uzbekvoice/check_voice.py
@@ -71,10 +71,11 @@ async def ask_action_handler(call: CallbackQuery, state: FSMContext):
     elif call_data == 'skip':
         await skip_voice(voice_id, chat_id)
         await call.message.delete()
-    else:
-        await send_message(call.chat.id, 'ask-to-confirm', markup=confirm_action_markup)
-        await call.message.delete_reply_markup()
-        await send_voice_vote(voice_id, call_data, chat_id)
+
+    elif call_data in ['accept', 'reject']:
+        await state.update_data(call_data=call_data)
+        await edit_reply_markup(chat_id, message_id, confirm_action_markup)
+        await AskUserAction.confirm_action.set()
 
     # If there are no more voice to check, get new list of text
     if list_number == 4:
@@ -88,13 +89,28 @@ async def ask_action_handler(call: CallbackQuery, state: FSMContext):
     else:
         await state.update_data(list_number=list_number + 1)
 
-    await ask_to_check_voice(chat_id, state)
 
-
-@dp.callback_query_handler(state=AskUserAction.confirm_action, text=['accept', 'reject'])
+@dp.callback_query_handler(state=AskUserAction.confirm_action, text=['confirm', 'back'])
 async def ask_action_handler(call: CallbackQuery, state: FSMContext):
-    await call.message.delete_reply_markup()
-    await send_voice_vote(voice_id, call_data, chat_id)
+    call_data = str(call.data)
+    chat_id = call.message.chat.id
+    message_id = call.message.message_id
+
+    await call.answer()
+
+    data = await state.get_data()
+    list_number = data['list_number']
+    voices_info = data['voices_info']
+    voice_id = voices_info[list_number]['id']
+    if call_data == 'confirm':
+        await call.message.delete_reply_markup()
+        await send_voice_vote(voice_id, data['call_data'], chat_id)
+        await ask_to_check_voice(chat_id, state)
+
+    elif call_data == 'back':
+        await edit_reply_markup(chat_id, message_id, yes_no_markup)
+        await AskUserAction.ask_action.set()
+        return
 
 
 # Handler that receives action on pressed report inline button
